fixes new config-reloader
