+++ internal/config-reloader/Dockerfile
@@ -9,7 +9,7 @@ ENV USER_UID=1001 \
     USER_NAME=config-reloader
 
 # install operator binary
-COPY bin/config-reloader /usr/local/bin/config-reloader
+COPY bin/config-reloader-${ARCH} /usr/local/bin/config-reloader
 
 ENTRYPOINT ["/usr/local/bin/config-reloader"]
