+++ internal/config-reloader/README.md
@@ -1,0 +1,6 @@
+### Config-reloader
+
+
+ It's alternative version of `prometheus-config-reloader`.
+ The main difference is ability to read secret directly from kubernetes and write it to local file system.
+ It should speed-up config reloading process and makes it more predictable.
