+++ internal/config/config.go
@@ -34,7 +34,7 @@ type BaseOperatorConf struct {
 	// enables custom config reloader for vmauth and vmagent,
 	// it should speed-up config reloading process.
 	UseCustomConfigReloader   bool   `default:"false"`
-	CustomConfigReloaderImage string `default:"victoriametrics/operator:config-reloader.0.1.0-amd64"`
+	CustomConfigReloaderImage string `default:"victoriametrics/operator:config-reloader-0.1.0"`
 	PSPAutoCreateEnabled      bool   `default:"true"`
 	VMAlertDefault            struct {
 		Image               string `default:"victoriametrics/vmalert"`
